+++ src/pytorch_metric_learning/losses/soft_triple_loss.py
@@ -23,7 +23,7 @@ class SoftTripleLoss(BaseMetricLossFunction):
         self.add_to_recordable_attributes(list_of_names=["same_class_center_sim", "diff_class_center_sim"], is_stat=True)
 
     def compute_loss(self, embeddings, labels, indices_tuple):
-        miner_weights = lmu.convert_to_weights(indices_tuple, labels)
+        miner_weights = lmu.convert_to_weights(indices_tuple, labels, dtype=embeddings.dtype)
         centers = F.normalize(self.fc, p=2, dim=0) if self.normalize_embeddings else self.fc
         sim_to_centers = torch.matmul(embeddings, centers)
         sim_to_centers = sim_to_centers.view(-1, self.num_classes, self.centers_per_class)
